5.0.2

+++ src/main/java/it/smartcommunitylab/aac/core/service/ConfigurableProviderEntityService.java
@@ -19,50 +19,64 @@ package it.smartcommunitylab.aac.core.service;
 import it.smartcommunitylab.aac.common.NoSuchProviderException;
 import it.smartcommunitylab.aac.core.persistence.ProviderEntity;
 import it.smartcommunitylab.aac.core.persistence.ProviderEntityRepository;
-import java.io.Serializable;
 import java.util.List;
-import java.util.Map;
-import org.apache.commons.lang3.RandomStringUtils;
+import java.util.UUID;
+import org.springframework.lang.Nullable;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
+import org.springframework.util.StringUtils;
 
-public class ConfigurableProviderEntityService<P extends ProviderEntity> implements ProviderEntityService<P> {
+@Service
+@Transactional
+public class ConfigurableProviderEntityService {
 
-    private final ProviderEntityRepository<P> providerRepository;
+    protected final ProviderEntityRepository providerRepository;
 
-    public ConfigurableProviderEntityService(ProviderEntityRepository<P> providerRepository) {
+    public ConfigurableProviderEntityService(ProviderEntityRepository providerRepository) {
         Assert.notNull(providerRepository, "provider repository is mandatory");
+
         this.providerRepository = providerRepository;
     }
 
     @Transactional(readOnly = true)
-    public List<P> listProviders() {
-        return providerRepository.findAll();
+    public List<ProviderEntity> listProviders(String type) {
+        return providerRepository.findByType(type);
     }
 
     @Transactional(readOnly = true)
-    public List<P> listProvidersByAuthority(String authority) {
-        return providerRepository.findByAuthority(authority);
-    }
-
-    @Transactional(readOnly = true)
-    public List<P> listProvidersByRealm(String realm) {
+    public List<ProviderEntity> listProvidersByRealm(String realm) {
         return providerRepository.findByRealm(realm);
     }
 
     @Transactional(readOnly = true)
-    public List<P> listProvidersByAuthorityAndRealm(String authority, String realm) {
-        return providerRepository.findByAuthorityAndRealm(authority, realm);
+    public List<ProviderEntity> listProvidersByRealm(String type, String realm) {
+        return providerRepository.findByTypeAndRealm(type, realm);
     }
 
     @Transactional(readOnly = true)
-    public P findProvider(String providerId) {
-        return providerRepository.findByProvider(providerId);
+    public List<ProviderEntity> listProvidersByAuthority(String type, String authority) {
+        return providerRepository.findByTypeAndAuthority(type, authority);
     }
 
     @Transactional(readOnly = true)
-    public P getProvider(String providerId) throws NoSuchProviderException {
-        P p = providerRepository.findByProvider(providerId);
+    public List<ProviderEntity> listProvidersByAuthorityAndRealm(String type, String authority, String realm) {
+        return providerRepository.findByTypeAndAuthorityAndRealm(type, authority, realm);
+    }
+
+    @Transactional(readOnly = true)
+    public ProviderEntity findProvider(String type, String providerId) {
+        ProviderEntity p = providerRepository.findByProvider(providerId);
+        if (p == null || !type.equals(p.getType())) {
+            return null;
+        }
+
+        return p;
+    }
+
+    @Transactional(readOnly = true)
+    public ProviderEntity getProvider(String type, String providerId) throws NoSuchProviderException {
+        ProviderEntity p = findProvider(type, providerId);
         if (p == null) {
             throw new NoSuchProviderException();
         }
@@ -70,46 +84,66 @@ public class ConfigurableProviderEntityService<P extends ProviderEntity> impleme
         return p;
     }
 
-    public P saveProvider(String providerId, P reg, Map<String, Serializable> configurationMap) {
-        P p = reg;
+    //TODO drop nullable id, we shoulds always receive it
+    public ProviderEntity saveProvider(String type, @Nullable String providerId, ProviderEntity reg) {
+        if (reg == null) {
+            throw new IllegalArgumentException();
+        }
+
+        if (StringUtils.hasText(reg.getType()) && !type.equals(reg.getType())) {
+            throw new IllegalArgumentException();
+        }
 
         if (providerId == null) {
             providerId = generateId();
         }
 
-        // always set id
-        p.setProvider(providerId);
+        ProviderEntity p = providerRepository.findByProvider(providerId);
+        if (p == null) {
+            p = new ProviderEntity();
+            p.setProvider(providerId);
+            p.setType(type);
+        } else {
+            if (!type.equals(p.getType())) {
+                throw new IllegalArgumentException();
+            }
+        }
 
-        // set configMap
-        p.setConfigurationMap(configurationMap);
+        //set base
+        p.setAuthority(reg.getAuthority());
+        p.setRealm(reg.getRealm());
+
+        p.setName(reg.getName());
+        p.setTitleMap(reg.getTitleMap());
+        p.setDescriptionMap(reg.getDescriptionMap());
+
+        //set config
+        int version = reg.getVersion() != null ? reg.getVersion().intValue() : 0;
+        p.setSettingsMap(reg.getSettingsMap());
+        p.setConfigurationMap(reg.getConfigurationMap());
+        p.setVersion(version);
+
+        //status
+        boolean enabled = reg.getEnabled() != null ? reg.getEnabled().booleanValue() : false;
+        p.setEnabled(enabled);
 
         // save and flush to ensure id is properly persisted
         p = providerRepository.saveAndFlush(p);
         return p;
     }
 
-    public void deleteProvider(String providerId) {
-        P p = providerRepository.findByProvider(providerId);
-        if (p != null) {
+    public void deleteProvider(String type, String providerId) {
+        ProviderEntity p = providerRepository.findByProvider(providerId);
+        if (p != null && type.equals(p.getType())) {
             providerRepository.delete(p);
         }
     }
 
     /*
      * Helpers
+     * TODO drop from here
      */
     private String generateId() {
-        // generate a random Id with fallback check
-        // TODO evaluate UUID replacement
-        // generate random
-        String id = RandomStringUtils.randomAlphanumeric(8);
-
-        P pe = findProvider(id);
-        if (pe != null) {
-            // re generate longer
-            id = RandomStringUtils.randomAlphanumeric(10);
-        }
-
-        return id;
+        return UUID.randomUUID().toString();
     }
 }
